requirements.txt
openai==1.1.1
anthropic>=0.16.0
python-dotenv==1.0.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
faiss-cpu==1.7.4
numpy==1.24.3
Pillow==10.0.0
pyautogui==0.9.54
pygetwindow==0.0.9
pywinauto==0.6.8
psutil==5.9.5
pdf2image==1.16.3
PyMuPDF==1.23.3